Update generated files with build 70488

+++ chats/item/sendactivitynotification/send_activity_notification_request_body.go
@@ -12,7 +12,7 @@ type SendActivityNotificationRequestBody struct {
     // Stores additional data not described in the OpenAPI description found when deserializing. Can be used for serialization as well.
     additionalData map[string]interface{};
     // 
-    chainId *int64;
+    chainId *int32;
     // 
     previewText i535684e11b5500196ecb4b5c6634e0651fe2c2f78b6cd0fbe097d3c9029ae7bc.ItemBodyable;
     // 
@@ -50,7 +50,7 @@ func (m *SendActivityNotificationRequestBody) GetAdditionalData()(map[string]int
     }
 }
 // GetChainId gets the chainId property value. 
-func (m *SendActivityNotificationRequestBody) GetChainId()(*int64) {
+func (m *SendActivityNotificationRequestBody) GetChainId()(*int32) {
     if m == nil {
         return nil
     } else {
@@ -71,7 +71,7 @@ func (m *SendActivityNotificationRequestBody) GetFieldDeserializers()(map[string
         return nil
     }
     res["chainId"] = func (o interface{}, n i04eb5309aeaafadd28374d79c8471df9b267510b4dc2e3144c378c50f6fd7b55.ParseNode) error {
-        val, err := n.GetInt64Value()
+        val, err := n.GetInt32Value()
         if err != nil {
             return err
         }
@@ -170,7 +170,7 @@ func (m *SendActivityNotificationRequestBody) Serialize(writer i04eb5309aeaafadd
         }
     }
     {
-        err := writer.WriteInt64Value("chainId", m.GetChainId())
+        err := writer.WriteInt32Value("chainId", m.GetChainId())
         if err != nil {
             return err
         }
@@ -224,7 +224,7 @@ func (m *SendActivityNotificationRequestBody) SetAdditionalData(value map[string
     }
 }
 // SetChainId sets the chainId property value. 
-func (m *SendActivityNotificationRequestBody) SetChainId(value *int64)() {
+func (m *SendActivityNotificationRequestBody) SetChainId(value *int32)() {
     if m != nil {
         m.chainId = value
     }
