+++ deviceappmanagement/mobileapps/getmobileappcountwithstatus/get_mobile_app_count_with_status_responseable.go
@@ -8,6 +8,6 @@ import (
 type GetMobileAppCountWithStatusResponseable interface {
     i04eb5309aeaafadd28374d79c8471df9b267510b4dc2e3144c378c50f6fd7b55.AdditionalDataHolder
     i04eb5309aeaafadd28374d79c8471df9b267510b4dc2e3144c378c50f6fd7b55.Parsable
-    GetValue()(*int64)
-    SetValue(value *int64)()
+    GetValue()(*int32)
+    SetValue(value *int32)()
 }
