+++ models/microsoft/graph/application_sign_in_summaryable.go
@@ -9,11 +9,11 @@ type ApplicationSignInSummaryable interface {
     Entityable
     i04eb5309aeaafadd28374d79c8471df9b267510b4dc2e3144c378c50f6fd7b55.Parsable
     GetAppDisplayName()(*string)
-    GetFailedSignInCount()(*int64)
-    GetSuccessfulSignInCount()(*int64)
+    GetFailedSignInCount()(*int32)
+    GetSuccessfulSignInCount()(*int32)
     GetSuccessPercentage()(*float64)
     SetAppDisplayName(value *string)()
-    SetFailedSignInCount(value *int64)()
-    SetSuccessfulSignInCount(value *int64)()
+    SetFailedSignInCount(value *int32)()
+    SetSuccessfulSignInCount(value *int32)()
     SetSuccessPercentage(value *float64)()
 }
