+++ models/microsoft/graph/certificate_connector_health_metric_value.go
@@ -12,9 +12,9 @@ type CertificateConnectorHealthMetricValue struct {
     // Timestamp for this metric data-point.
     dateTime *i336074805fc853987abe6f7fe3ad97a6a6f3077a16391fec744f671a015fbd7e.Time;
     // Count of failed requests/operations.
-    failureCount *int64;
+    failureCount *int32;
     // Count of successful requests/operations.
-    successCount *int64;
+    successCount *int32;
 }
 // NewCertificateConnectorHealthMetricValue instantiates a new certificateConnectorHealthMetricValue and sets the default values.
 func NewCertificateConnectorHealthMetricValue()(*CertificateConnectorHealthMetricValue) {
@@ -44,7 +44,7 @@ func (m *CertificateConnectorHealthMetricValue) GetDateTime()(*i336074805fc85398
     }
 }
 // GetFailureCount gets the failureCount property value. Count of failed requests/operations.
-func (m *CertificateConnectorHealthMetricValue) GetFailureCount()(*int64) {
+func (m *CertificateConnectorHealthMetricValue) GetFailureCount()(*int32) {
     if m == nil {
         return nil
     } else {
@@ -65,7 +65,7 @@ func (m *CertificateConnectorHealthMetricValue) GetFieldDeserializers()(map[stri
         return nil
     }
     res["failureCount"] = func (o interface{}, n i04eb5309aeaafadd28374d79c8471df9b267510b4dc2e3144c378c50f6fd7b55.ParseNode) error {
-        val, err := n.GetInt64Value()
+        val, err := n.GetInt32Value()
         if err != nil {
             return err
         }
@@ -75,7 +75,7 @@ func (m *CertificateConnectorHealthMetricValue) GetFieldDeserializers()(map[stri
         return nil
     }
     res["successCount"] = func (o interface{}, n i04eb5309aeaafadd28374d79c8471df9b267510b4dc2e3144c378c50f6fd7b55.ParseNode) error {
-        val, err := n.GetInt64Value()
+        val, err := n.GetInt32Value()
         if err != nil {
             return err
         }
@@ -87,7 +87,7 @@ func (m *CertificateConnectorHealthMetricValue) GetFieldDeserializers()(map[stri
     return res
 }
 // GetSuccessCount gets the successCount property value. Count of successful requests/operations.
-func (m *CertificateConnectorHealthMetricValue) GetSuccessCount()(*int64) {
+func (m *CertificateConnectorHealthMetricValue) GetSuccessCount()(*int32) {
     if m == nil {
         return nil
     } else {
@@ -106,13 +106,13 @@ func (m *CertificateConnectorHealthMetricValue) Serialize(writer i04eb5309aeaafa
         }
     }
     {
-        err := writer.WriteInt64Value("failureCount", m.GetFailureCount())
+        err := writer.WriteInt32Value("failureCount", m.GetFailureCount())
         if err != nil {
             return err
         }
     }
     {
-        err := writer.WriteInt64Value("successCount", m.GetSuccessCount())
+        err := writer.WriteInt32Value("successCount", m.GetSuccessCount())
         if err != nil {
             return err
         }
@@ -138,13 +138,13 @@ func (m *CertificateConnectorHealthMetricValue) SetDateTime(value *i336074805fc8
     }
 }
 // SetFailureCount sets the failureCount property value. Count of failed requests/operations.
-func (m *CertificateConnectorHealthMetricValue) SetFailureCount(value *int64)() {
+func (m *CertificateConnectorHealthMetricValue) SetFailureCount(value *int32)() {
     if m != nil {
         m.failureCount = value
     }
 }
 // SetSuccessCount sets the successCount property value. Count of successful requests/operations.
-func (m *CertificateConnectorHealthMetricValue) SetSuccessCount(value *int64)() {
+func (m *CertificateConnectorHealthMetricValue) SetSuccessCount(value *int32)() {
     if m != nil {
         m.successCount = value
     }
