+++ models/microsoft/graph/certificate_connector_setting.go
@@ -18,7 +18,7 @@ type CertificateConnectorSetting struct {
     // Last time certificate connector connected
     lastConnectorConnectionTime *i336074805fc853987abe6f7fe3ad97a6a6f3077a16391fec744f671a015fbd7e.Time;
     // Version of last uploaded certificate connector
-    lastUploadVersion *int64;
+    lastUploadVersion *int32;
     // Certificate connector status
     status *int32;
 }
@@ -109,7 +109,7 @@ func (m *CertificateConnectorSetting) GetFieldDeserializers()(map[string]func(in
         return nil
     }
     res["lastUploadVersion"] = func (o interface{}, n i04eb5309aeaafadd28374d79c8471df9b267510b4dc2e3144c378c50f6fd7b55.ParseNode) error {
-        val, err := n.GetInt64Value()
+        val, err := n.GetInt32Value()
         if err != nil {
             return err
         }
@@ -139,7 +139,7 @@ func (m *CertificateConnectorSetting) GetLastConnectorConnectionTime()(*i3360748
     }
 }
 // GetLastUploadVersion gets the lastUploadVersion property value. Version of last uploaded certificate connector
-func (m *CertificateConnectorSetting) GetLastUploadVersion()(*int64) {
+func (m *CertificateConnectorSetting) GetLastUploadVersion()(*int32) {
     if m == nil {
         return nil
     } else {
@@ -184,7 +184,7 @@ func (m *CertificateConnectorSetting) Serialize(writer i04eb5309aeaafadd28374d79
         }
     }
     {
-        err := writer.WriteInt64Value("lastUploadVersion", m.GetLastUploadVersion())
+        err := writer.WriteInt32Value("lastUploadVersion", m.GetLastUploadVersion())
         if err != nil {
             return err
         }
@@ -234,7 +234,7 @@ func (m *CertificateConnectorSetting) SetLastConnectorConnectionTime(value *i336
     }
 }
 // SetLastUploadVersion sets the lastUploadVersion property value. Version of last uploaded certificate connector
-func (m *CertificateConnectorSetting) SetLastUploadVersion(value *int64)() {
+func (m *CertificateConnectorSetting) SetLastUploadVersion(value *int32)() {
     if m != nil {
         m.lastUploadVersion = value
     }
