+++ models/microsoft/graph/credential_usage_summaryable.go
@@ -9,11 +9,11 @@ type CredentialUsageSummaryable interface {
     Entityable
     i04eb5309aeaafadd28374d79c8471df9b267510b4dc2e3144c378c50f6fd7b55.Parsable
     GetAuthMethod()(*UsageAuthMethod)
-    GetFailureActivityCount()(*int64)
+    GetFailureActivityCount()(*int32)
     GetFeature()(*FeatureType)
-    GetSuccessfulActivityCount()(*int64)
+    GetSuccessfulActivityCount()(*int32)
     SetAuthMethod(value *UsageAuthMethod)()
-    SetFailureActivityCount(value *int64)()
+    SetFailureActivityCount(value *int32)()
     SetFeature(value *FeatureType)()
-    SetSuccessfulActivityCount(value *int64)()
+    SetSuccessfulActivityCount(value *int32)()
 }
