+++ models/microsoft/graph/credential_user_registration_countable.go
@@ -8,8 +8,8 @@ import (
 type CredentialUserRegistrationCountable interface {
     Entityable
     i04eb5309aeaafadd28374d79c8471df9b267510b4dc2e3144c378c50f6fd7b55.Parsable
-    GetTotalUserCount()(*int64)
+    GetTotalUserCount()(*int32)
     GetUserRegistrationCounts()([]UserRegistrationCountable)
-    SetTotalUserCount(value *int64)()
+    SetTotalUserCount(value *int32)()
     SetUserRegistrationCounts(value []UserRegistrationCountable)()
 }
