+++ models/microsoft/graph/externalconnectors/connection_quotaable.go
@@ -9,6 +9,6 @@ import (
 type ConnectionQuotaable interface {
     i535684e11b5500196ecb4b5c6634e0651fe2c2f78b6cd0fbe097d3c9029ae7bc.Entityable
     i04eb5309aeaafadd28374d79c8471df9b267510b4dc2e3144c378c50f6fd7b55.Parsable
-    GetItemsRemaining()(*int64)
-    SetItemsRemaining(value *int64)()
+    GetItemsRemaining()(*int32)
+    SetItemsRemaining(value *int32)()
 }
